updated enemy prefab

+++ HorseHouse-GameJam/Assets/Dev/Scripts/Managers/EventManager.cs
@@ -28,8 +28,8 @@ public class EventManager : MonoBehaviour
     public void TakeDamage(int health) { takeDamage?.Invoke(health); }
     public event Action<int> updateHealth;
     public void UpdateHealth(int health) { updateHealth?.Invoke(health); }
-    public event Action<float,Vector2> enemyDeath;
-    public void EnemyDeath(float progress, Vector2 position) { enemyDeath?.Invoke(progress, position); }
+    public event Action<int,Vector2> enemyDeath;
+    public void EnemyDeath(int progress, Vector2 position) { enemyDeath?.Invoke(progress, position); }
     #endregion
 
 }
